+++ packages/wise_zitadel_login/lib/wise_zitadel_login.dart
@@ -3,6 +3,7 @@ library wise_zitadel_login;
 
 export 'src/options.dart';
 export 'src/providers/wise_zitadel_provider.dart';
+export 'src/repository/auth_repository.dart' show WiseZitadelAuthenticator;
 export 'src/routes/wise_login_screen_route.dart';
 export 'src/screens/wise_login_screen.dart';
 export 'src/types/zitadel_login_type.dart';